requirements.txt
python-binance==1.0.19
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
python-dotenv==1.0.0
ccxt      # 用于获取加密货币数据 